app/requirements.txt
GitPython==3.1.24
pythonping==1.1.0
prometheus-client==0.11.0